Uwzględnienie kolejnych uwag.
